OK; added EdgeVerticleCore instead
